cambios antes de pull con yami
