+++ api/src/utils/calculeScore.js
@@ -1,0 +1,17 @@
+
+
+const calculeScore = (arrayScores) => {
+    let sumScore = 0;
+    arrayScores.map((r) => {
+  
+      sumScore = sumScore + parseInt(r.score);
+    });
+  
+    const totalScore = sumScore / arrayScores.length;
+  
+    return totalScore;
+  };
+
+
+ module.exports = { calculeScore }
+
